# openapi.yml
openapi: 3.0.0
info:
    version: 1.0.0
    title: Sample API
    description: A sample API to illustrate OpenAPI concepts
servers:
    - url: "http://localhost:8080/"
paths:
...
    /user:
        get:
            tags:
            - user
            summary: returns all users
            description: returns all users
            operationId: getAllUsers
            responses:
                "200":
                    description: returns all users
                    content: 
                        application/json:
                            schema:
                             $ref: "#/components/schemas/User"
                "500":
                    description: internal error
        post:
            tags:
            - user
            summary: Create user
            description: This can only be done by the logged in user.
            operationId: createUser
           
           
            requestBody:
                description: Created user object
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
                required: true
            
            
            responses:
                default:
                    description: successful operation
                    content: {}
            x-codegen-request-body-name: body
...
    /user/userId:
     delete:
      tags:
      - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
      - name: username
        in: path
        description: The name that needs to be deleted
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid username supplied
          content: {}
        404:
          description: User not found
          content: {}
components:
    schemas:
        User:
            required:
                - nickname
                - surname
                - name
            type: object
            properties:
                id :
                    type : number
                name : 
                    type : string
                surname: 
                    type : string
                nickname:
                    type: string
                age:
                    type: number
                email:
                    type: string
                balance: 
                    type: number
                wins:
                    type: number
                losses:
                    type: number
                moneywon:
                    type: number
                creationdate:
                    type: number
...